addressing suggestions
