--- api/fields_parser.py
@@ -1,9 +1,0 @@
-def custom_fields_parser(root_key, key_path, val):
-    """ consumes values like ("a",["foo"],["baz,hello","world"])
-        returns (a, {"foo": {"baz": True, "hello": True, "world": True}}}, is_deep_object)
-    """
-    root = {key_path[0]: {}}
-    for v in val:
-        for fields in v.split(","):
-            root[key_path[0]][fields] = True
-    return (root_key, [root], True)
