+++ api/parsing.py
@@ -2,7 +2,16 @@ import re
 
 from connexion.decorators.uri_parsing import OpenAPIURIParser
 
-from api.fields_parser import custom_fields_parser
+
+def custom_fields_parser(root_key, key_path, val):
+    """ consumes values like ("a",["foo"],["baz,hello","world"])
+        returns (a, {"foo": {"baz": True, "hello": True, "world": True}}}, is_deep_object)
+    """
+    root = {key_path[0]: {}}
+    for v in val:
+        for fields in v.split(","):
+            root[key_path[0]][fields] = True
+    return (root_key, [root], True)
 
 
 class customURIParser(OpenAPIURIParser):
